refactoring & cleanup (01)
