+++ src/Main.cpp
@@ -4,7 +4,7 @@
 
 #include <tnt/store/TripleStore.hpp>
 #include <tnt/config/Config.cpp>
-#include <tnt/http/AtomicTripleStore.hpp>
+#include <tnt/store/AtomicTripleStore.hpp>
 #include <tnt/http/SPARQLEndpoint.hpp>
