+++ src/tnt/http/JSONSparqlResultsHeader.hpp
